fix: Use parseFloat for resend delay to support 3-minute test option
parseInt was truncating 0.05 to 0, breaking the test delay option.

+++ app/(mact)/outreach/new/page.tsx
@@ -892,7 +892,7 @@ export default function NewCampaignPage() {
                       <Label>Wait before resending</Label>
                       <Select
                         value={resendDelayHours.toString()}
-                        onValueChange={(v) => setResendDelayHours(parseInt(v))}
+                        onValueChange={(v) => setResendDelayHours(parseFloat(v))}
                       >
                         <SelectTrigger className="w-48">
                           <SelectValue />
